restructuring directories to build solution structure
create solution, xunit testing, reproducible program structure

add building/running instructions for project and tests

hello world test with solution setup
