+++ mongoCluster.Tests/ProgramTest.cs
@@ -1,0 +1,15 @@
+using System;
+using Xunit;
+
+namespace mongoCluster.Tests
+{
+    public class UnitTest1
+    {
+        [Fact]
+        public void SolutionStructureCorrect_ConsolePrintSuccess()
+        {
+		Console.WriteLine("Program Test Correctly Referenced");
+
+        }
+    }
+}
