+++ mongoCluster/NLog.config
@@ -1,0 +1,17 @@
+<?xml version="1.0" encoding="utf-8" ?>
+<nlog xmlns="http://www.nlog-project.org/schemas/NLog.xsd"
+      xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+      xsi:schemaLocation="http://www.nlog-project.org/schemas/NLog.xsd NLog.xsd"
+      autoReload="true"
+      throwExceptions="false"
+      internalLogLevel="Off" internalLogFile="c:\temp\nlog-internal.log">
+
+    <targets>
+      <target name="console" xsi:type="Console" layout="${longdate}|${message}"/>
+    </targets>
+
+    <rules>
+      <logger name="*" minlevel="Debug" writeTo="console" />
+    </rules>
+   
+</nlog>
